requirements.txt
slack-bolt>=1.19,<2.0
python-dotenv>=1.0
requests>=2.32
